testing in progress 1

+++ src/controllers/fiwareController.js
@@ -1,5 +1,6 @@
 const axios = require('axios');
 
+
 const SERVER_IP  = '46.17.108.45';
 const ORION_PORT = '1026';
 const API_KEY    = 'asd12345';
@@ -49,57 +50,41 @@ exports.getOrionEntities = async function (req, res) {
     }
 };
 
-// create Orion Entity
+// create Orion Ambiente Entity
 exports.createOrionEntity = async function(req,res) {
 
     // params must come from req.body
 
-    let eName = "new_entity_2";
-    let eId   = "new:002";
-    let eType  = "Ambiente";
-    let eDescription = "A new entity for y purpose";
+   let { id, type, description, temperature, humidity, acidity } = req.body;
 
-    let service = "sensor";
-    let ServicePath = "/";
+   if(!id || !type || !description || !temperature || !humidity || !acidity) {
+         return res.status(400).send({error: "Missing parameters"});
+    }
     
-    let data = JSON.stringify({
-        "id":   eId,
-        "type": eType,
-        "description": {
-            "value": eDescription,
-            "type":  "Text"
-        },
-        // ENTITY VALUES _ Structure
-        "online": {
-            "value": true,
-            "type": "Boolean"
-        },
-        "temperature": {
-            "value": 24,
-            "type": "Number"
-        },
-        "humidity": {
-            "value": 56,
-            "type": "Number"
-        },
-        "pressure": {
-            "value": 1,
-            "type": "Number"
-        },
-        
-    });
+    // let service = "sensor";
+    // let servicePath = "/";
+    
+    let data = {
+        "id":   id,
+        "type": type,
+        "description": description,
+        "temperature": temperature,
+        "humidity": humidity,
+        "accidity": acidity
+    };
 
     let config = {
         method: 'post',
         url: 'http://'+ SERVER_IP +':'+ ORION_PORT +'/v2/entities',
         headers: { 
-            'Fiware-Service': service, 
-            'Fiware-ServicePath': ServicePath, 
-            'X-Auth-Token': API_KEY, 
-            'Content-Type': 'application/json'
+            // 'Fiware-Service': service, 
+            // 'Fiware-ServicePath': servicePath, 
+            // 'X-Auth-Token': API_KEY, 
+            'Content-Type': 'application/json',
+            'Connection': 'keep-alive'
         },
-            timeout: 2000,
-            data : data
+        timeout: 2000,
+        data : data
     };
 
     try {
@@ -115,26 +100,23 @@ exports.createOrionEntity = async function(req,res) {
 // Delete Orion Entity
 exports.deleteEntity = async function(req,res) {
     
-    // entity data
-    // let entityId = req.body.entityId;
-    let service     = "sensor";
-    let servicePath = "/";
-    let entityId    = "new:002";
+    let { id } = req.body;
+    console.log("el id es esta " + id);
+
+    if(!id) {
+        return res.status(400).send({error: "Missing parameters"});
+    }
 
     let config = {
         method: 'delete',
-        url: 'http://'+ SERVER_IP +':'+ ORION_PORT +'/v2/entities/' + entityId,
-        headers: { 
-            'Fiware-Service': service, 
-            'Fiware-ServicePath': servicePath, 
-            'X-Auth-Token': API_KEY
-        },
+        url: 'http://'+ SERVER_IP +':'+ ORION_PORT +'/v2/entities/' + id,
+        headers: {},
         timeout: 2000,
     };
 
     try {
         let response = await axios(config);
-        return res.status(201).send(response.data);
+        return res.status(204).send(response.data);
     } catch (err) {
         console.log(err)
         return res.send(err);
